Разместил на главной кнопку "войти"
